refactor: Move PasskeyAuthFilter to filters pkg
